Updates publications.pl to allow boolean search sting in keywords
So now we can limit the articles returned from the lab.bib file by
generating a search string like 'lab NOT hmmer' and it will return all
articles that have the keyword lab but not the keyword hmmer. AND, NOT
and OR are all allowed terms and the search can be broken down into sub
queries with ().

git-svn-id: https://svn.janelia.org/eddylab/eddys/labbib/trunk@240 d9ddfb79-221d-0410-8357-8f9f680106d2