Allows publications.pl to accept different .bib files
Now when calling publications.pl you can pass it the name of a .bib file
and it will use that instead of lab.bib. For example:

publications.pl master.bib > output.html

will take the contents of master.bib and convert it into html to be
placed in the output.html file.

git-svn-id: https://svn.janelia.org/eddylab/eddys/labbib/trunk@242 d9ddfb79-221d-0410-8357-8f9f680106d2