Adds option to specify alternative output templates.
Since this system processes .bib files for multiples sites, it needs to
be able to modify the output to suit each case. This allows us to
specify a template on the command line for each of those cases. The
default template will produce a basic unordered list. Additional
templates should be located in the directory from which the script is
called.

git-svn-id: https://svn.janelia.org/eddylab/eddys/labbib/trunk@244 d9ddfb79-221d-0410-8357-8f9f680106d2