pmid2bib replaced with a much better script.
pmid2bib <pmid> returns a BibTeX-formatted entry, based on parsing
NCBI XML. This is more robust (by far) than how I've done it in the
past, and it just required understanding how to use a good XML parser
(xml.etree.ElementTree) and the current NCBI PubMed DTD
(https://dtd.nlm.nih.gov/ncbi/pubmed/out/pubmed_230101.dtd).

Deleted previous pmid2bib, which didn't really work well.

Also deleted pubmed2bib, which hasn't been working for a while; this
was based on cut-and-pasting the displayed PubMed abstract page from
the browser, and they change that format (and make it more and more
complicated) too often for this to be a viable strategy. Much better
to parse the XML.

Was motivated to do this after doing the HHMI renewal, where I ended
up having to hand-type a bunch of BibTeX entries because nothing was
working when I needed it, of course.